new users folder structure added but still old stats logic
