+++ backend/users/Dockerfile
@@ -8,9 +8,9 @@ RUN npm clean-install
 
 COPY . .
 
-RUN npx prisma generate
+RUN npx prisma generate 
 
-RUN npx prisma db push
+RUN npx prisma db push --accept-data-loss
 
 
 RUN npm install -g ts-node
