+++ backend/users/prisma/schema.prisma
@@ -8,21 +8,24 @@ datasource db {
 }
 
 model User {
-    id                     String            @id @default(cuid())
-    username               String            @unique
-    email                  String            @unique
-    firstname              String
-    password               String
-    avatarUrl              String            @default("/assets/img/avatar.jpg")
-    isOnline               Boolean           @default(false)
-    createdAt              DateTime          @default(now())
+    id          String   @id @default(cuid())
+    username    String   @unique
+    email       String   @unique
+    firstname   String
+    password    String
+    avatarUrl   String   @default("/assets/img/avatar.jpg")
+    isOnline    Boolean  @default(false)
+    createdAt   DateTime @default(now())
     // pong stats
-    gamesPlayed            Int               @default(0)
-    gamesWon               Int               @default(0)
-    totalScore             Int               @default(0)
-    UserGameHistory        UserGameHistory[]
+    gamesPlayed Int      @default(0)
+    wins        Int      @default(0)
+    losses      Int      @default(0)
+    elo         Int      @default(1200) //initial elo
+
+    // relations
     sentFriendRequests     Friendship[]      @relation("sent_requests")
     receivedFriendRequests Friendship[]      @relation("received_requests")
+    gameHistory            UserGameHistory[] @relation("user_games")
 
     @@map("users")
 }
@@ -47,16 +50,20 @@ model Friendship {
 }
 
 model UserGameHistory {
-    id         String   @id @default(cuid())
-    userId     String
-    gameId     String // id of pong match in pong-db
+    id     String @id @default(cuid())
+    gameId String @unique // id of pong match in pong-db
+    //user relation
+    userId String
+    user   User   @relation("user_games", fields: [userId], references: [id])
+
+    //game data
     isWinner   Boolean
-    score      Int
     opponentId String?
+    userScore  Int // user score in the match
+    eloChange  Int // how much elo won or lost
     playedAt   DateTime @default(now())
 
-    user User @relation(fields: [userId], references: [id])
-
+    @@index([userId])
     @@map("user_game_history")
 }
