+++ backend/users/src/modules/stats.schema.ts
@@ -1,0 +1,16 @@
+import { Type, Static } from "@sinclair/typebox";
+
+// schema for the url params -> user id
+export const userIdParamsSchema = Type.Object({
+  id: Type.String({ format: "cuid" }), // we check that the id has cuid format
+});
+export type UserIdParams = Static<typeof userIdParamsSchema>;
+
+// body schema req for update stats
+export const updateStatsBodySchema = Type.Object({
+  gameId: Type.String(),
+  opponentId: Type.String({ format: "cuid" }),
+  isWinner: Type.Boolean(),
+  score: Type.Integer({ minimum: 0 }),
+});
+export type UpdateStatsBody = Static<typeof updateStatsBodySchema>;
